refactor(components): refactor descriptions /descriptions-item
