+++ packages/components/descriptions/__tests__/descriptions.spec.ts
@@ -1,7 +1,7 @@
 import { nextTick } from 'vue'
 import { mount } from '@vue/test-utils'
-import Descriptions from '../src/index.vue'
-import DescriptionsItem from '../src/description-item'
+import Descriptions from '../src/descriptions.vue'
+import DescriptionsItem from '../src/descriptions-item'
 
 const _mount = (template: string, data?: () => void, methods?: any) =>
   mount({
