+++ packages/components/descriptions/index.ts
@@ -1,7 +1,7 @@
 import { withInstall, withNoopInstall } from '@element-plus/utils/with-install'
 
-import Descriptions from './src/index.vue'
-import DescriptionsItem from './src/description-item'
+import Descriptions from './src/descriptions.vue'
+import DescriptionsItem from './src/descriptions-item'
 
 export const ElDescriptions = withInstall(Descriptions, {
   DescriptionsItem,
