+++ packages/components/descriptions/src/descriptions-cell.ts
@@ -4,10 +4,7 @@ import { getNormalizedProps } from '@element-plus/utils/vnode'
 import { elDescriptionsKey } from './token'
 
 import type { VNode } from 'vue'
-import type {
-  IDescriptionsInject,
-  IDescriptionsItemInject,
-} from './descriptions.type'
+import type { DescriptionsProps, IDescriptionsItemInject } from './descriptions'
 
 export default defineComponent({
   name: 'ElDescriptionsCell',
@@ -23,7 +20,7 @@ export default defineComponent({
     },
   },
   setup() {
-    const descriptions = inject(elDescriptionsKey, {} as IDescriptionsInject)
+    const descriptions = inject(elDescriptionsKey, {} as DescriptionsProps)
 
     return {
       descriptions,
