--- packages/components/descriptions/src/descriptions.type.ts
@@ -1,21 +1,0 @@
-import type { ComponentSize } from '@element-plus/utils/types'
-
-export interface IDescriptionsInject {
-  border: boolean
-  column: number
-  direction: 'horizontal' | 'vertical'
-  size: ComponentSize
-  title: string
-  extra: string
-}
-
-export interface IDescriptionsItemInject {
-  label: string
-  span: number
-  width: string | number
-  minWidth: string | number
-  align: string
-  labelAlign: string
-  className: string
-  labelClassName: string
-}
